// src/app/shared/components/dynamic-form.component/enums/image-video-document.enum.ts
export enum eFileAccept {
  IMAGE = 'image/png,image/jpeg',
  VIDEO = 'video/mp4',
  IMAGEANDVIDEO = 'image/png,image/jpeg,video/mp4',
  PDF = 'application/pdf',
  WORD = 'application/msword,application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  ALLDOCUMENT = 'application/pdf,application/msword,application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  ALL = 'image/png,image/jpeg,video/mp4,application/pdf,application/msword,application/vnd.openxmlformats-officedocument.wordprocessingml.document',
}